break circular dependence
