+++ src/applyBindingExport.js
@@ -1,0 +1,5 @@
+// re-export applyBinding to avoid circular dependence
+import applyBinding from './applyBinding';
+
+export default applyBinding;
+
